feat: update TW Marine dataset integration
Integrate TW Marine 5cls dataset into DETR inference pipeline. Updated
launch files, class definitions, and detection logic. Added support for
MarkerArray publishing and improved detection visualization. Refactored
distance and angle calculations for better accuracy.

+++ ros1_ws/src/detr_inference/src/detr_inference.py
@@ -138,8 +138,8 @@ class DetrInferenceNode:
                 box_bottom = []
                 for box in boxes:
                     x_min, y_min, x_max, y_max = box
-                    center_x = (x_min + x_max) / 2.0
-                    bottom_y = y_max
+                    center_x = int((x_min + x_max) / 2.0)
+                    bottom_y = int(y_max)
                     box_bottom.extend([center_x, bottom_y])
 
                 # Publish detections
